Move ModelForms from opengever.meeting to opengever.base
