+++ opengever/contact/browser/person.py
@@ -1,8 +1,8 @@
 from opengever.base.model import CONTENT_TITLE_LENGTH
 from opengever.contact import _
 from opengever.contact.models.person import Person
-from opengever.meeting.form import ModelAddForm
-from opengever.meeting.form import ModelEditForm
+from opengever.base.browser.modelforms import ModelAddForm
+from opengever.base.browser.modelforms import ModelEditForm
 from opengever.ogds.models import FIRSTNAME_LENGTH
 from opengever.ogds.models import LASTNAME_LENGTH
 from plone.directives import form
