+++ opengever/meeting/browser/members.py
@@ -1,8 +1,8 @@
 from Acquisition import aq_inner
 from Acquisition import aq_parent
 from opengever.meeting import _
-from opengever.meeting.form import ModelAddForm
-from opengever.meeting.form import ModelEditForm
+from opengever.base.browser.modelforms import ModelAddForm
+from opengever.base.browser.modelforms import ModelEditForm
 from opengever.meeting.model import Member
 from opengever.ogds.models import EMAIL_LENGTH
 from opengever.ogds.models import FIRSTNAME_LENGTH
