+++ opengever/meeting/browser/periods.py
@@ -5,8 +5,8 @@ from opengever.base.browser.wizard.interfaces import IWizardDataStorage
 from opengever.meeting import _
 from opengever.meeting import is_meeting_feature_enabled
 from opengever.meeting.committee import ICommittee
-from opengever.meeting.form import ModelAddForm
-from opengever.meeting.form import ModelEditForm
+from opengever.base.browser.modelforms import ModelAddForm
+from opengever.base.browser.modelforms import ModelEditForm
 from opengever.meeting.model import Period
 from plone.directives import form
 from plone.z3cform.layout import FormWrapper
